fix bug trigger bug and only show loading bar when appropriate
Fix a bug where after triggering a Shipment the Load Balancer Status
pane didn't show without either a hard refresh or navigating way and
then back to the Shipment.

Updated the Load Balancer Status pane to only show the indeterminate
loading bar when the description indicates that the status is temporal.